refactor, organize
