+++ src/components/HeadQuarters/head-quarters-reducer.js
@@ -2,7 +2,7 @@ import { fromJS, List } from 'immutable';
 import { types } from './head-quarters-actions';
 
 const INITIAL_STATE = fromJS({
-  numberOfHacks: 1,
+  numberOfHacks: 3,
   terminals: List(),
 });
