+++ src/components/PlayerBoard/player-board-reducer.js
@@ -5,7 +5,7 @@ const INITIAL_STATE = fromJS({
   colonists: 0,
   soylent: 0,
   credits: 0,
-  resources: {
+  distributedResources: {
     nanoTubes: 0,
     unobtanium: 0,
     energyPlankton: 0,
