+++ src/components/Terminal/image-paths.js
@@ -1,9 +1,9 @@
-import credits from '../assets/images/credits-green.png';
-import soylent from '../assets/images/soylent-green.png';
-import alienArtifact from '../assets/images/alien-artifact.png';
-import colonists from '../assets/images/colonists-green.png';
-import colonistsSoylent from '../assets/images/half-half-green.png';
-import alienDeath from '../assets/images/alien-and-death.png';
+import credits from '../assets/images/terminal-images/credits-green.png';
+import soylent from '../assets/images/terminal-images/soylent-green.png';
+import alienArtifact from '../assets/images/terminal-images/alien-artifact.png';
+import colonists from '../assets/images/terminal-images/colonists-green.png';
+import colonistsSoylent from '../assets/images/terminal-images/half-half-green.png';
+import alienDeath from '../assets/images/terminal-images/alien-and-death.png';
 
 const imagePaths = {
   'credits': {
