+++ src/components/assets/images/terminal-images/alien-and-death.png
