+++ src/components/assets/images/terminal-images/alien-artifact.png
