+++ src/components/assets/images/terminal-images/colonists-green.png
