+++ src/components/assets/images/terminal-images/credits-green.png
