+++ src/components/assets/images/terminal-images/half-half-green.png
