+++ src/components/assets/images/terminal-images/soylent-green.png
