+++ src/components/assets/images/terminal-images/terminal.png
